.foxe mirror
Use the repos as a mirror for the .foxe files. This is so that the files
can be referenced by the "raw" github url link, which permits cross
origin requests, and enables the extensions to be installed in the web
app directly from the app. The github "release" endpoints do not allow
cross origin requests.